# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: Произошла ошибка
        type: string
      success:
        example: false
        type: boolean
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: example@example.com
        type: string
      password_hash:
        example: "123456"
        minLength: 6
        type: string
    required:
    - email
    - password_hash
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: example@example.com
        type: string
      name:
        example: vasiliy1234
        minLength: 1
        type: string
      password_hash:
        example: "123456"
        minLength: 6
        type: string
    required:
    - email
    - name
    - password_hash
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        example: Операция выполнена успешно
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Support Team
  description: API для управления задачами, проектами и комментариями
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SaaS
  version: "1.0"
paths:
...
  /me:
    get:
      consumes:
      - application/json
      description: Возвращает данные авторизованного пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить информацию о текущем пользователе
      tags:
      - users
...
  /signin:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и возвращает JWT токен в cookie.
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Вход в систему
      tags:
      - auth
  /signout:
    post:
      consumes:
      - application/json
      description: Разлогинивает пользователя и удаляет куки авторизации.
      produces:
      - application/json
      responses:
        "200":
          description: успешный логаут
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "404":
          description: пользователь не найден
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Выход из системы
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя в системе.
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь зарегистрирован
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Неверный запрос или пользователь уже существует
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Введите "Bearer {token}" (без кавычек)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"